requirements_spaces.txt
tensorflow==2.15.0
scikit-learn==1.3.0
opencv-python-headless==4.8.0.76
gradio==4.44.1
scikit-image==0.21.0
Pillow==10.0.0
numpy==1.24.3
matplotlib==3.7.2
